TIP-1223: move interface in elasticsearch directory
