+++ src/Akeneo/Pim/Enrichment/Bundle/Elasticsearch/GetIndexableProductInterface.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Akeneo\Pim\Enrichment\Bundle\Elasticsearch;
+
+use Akeneo\Pim\Enrichment\Bundle\Elasticsearch\Model\ElasticsearchProductProjection;
+
+/**
+ * @author    Nicolas Marniesse <nicolas.marniesse@akeneo.com>
+ * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+interface GetIndexableProductInterface
+{
+    public function fromProductIdentifier(string $productIdentifier): ElasticsearchProductProjection;
+
+    public function fromProductIdentifiers(array $productIdentifiers): array;
+}
