+++ src/Akeneo/Pim/Enrichment/Bundle/Elasticsearch/Indexer/ProductIndexer.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Akeneo\Pim\Enrichment\Bundle\Elasticsearch\Indexer;
 
 use Akeneo\Pim\Enrichment\Component\Product\Connector\ReadModel\IndexableProduct;
-use Akeneo\Pim\Enrichment\Component\Product\Query\GetIndexableProductInterface;
+use Akeneo\Pim\Enrichment\Bundle\Elasticsearch\GetIndexableProductInterface;
 use Akeneo\Tool\Bundle\ElasticsearchBundle\Client;
 use Akeneo\Tool\Bundle\ElasticsearchBundle\Refresh;
 use Akeneo\Tool\Component\StorageUtils\Indexer\ProductIndexerInterface;
@@ -28,7 +28,7 @@ class ProductIndexer implements ProductIndexerInterface
     /** @var Client */
     private $productAndProductModelClient;
 
-    /** @var GetIndexableProductInterface */
+    /** @var \Akeneo\Pim\Enrichment\Bundle\Elasticsearch\GetIndexableProductInterface */
     private $getIndexableProduct;
 
     public function __construct(Client $productAndProductModelClient, GetIndexableProductInterface $getIndexableProduct)
